requirements.txt
Flask==3.1.3
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.2
cryptography==46.0.5
python-dotenv==1.2.2
requests==2.32.5
firebase-admin==7.2.0

numpy==1.26.4
pandas==2.3.3
scikit-learn==1.7.2
scipy==1.15.3

spacy==3.7.5
scispacy==0.6.2
negspacy==1.0.4
en-ner-bc5cdr-md==0.5.4
sentence-transformers==5.3.0
transformers==5.3.0
torch==2.10.0
faiss-cpu==1.13.2
